requirements.txt
pygame
numpy
scipy
neat-python
streamlit
pandas
